Refactor newserving->serving
