+++ Makefile
@@ -256,7 +256,7 @@ test_helpers:
 
 test_serving:
 	-mkdir coverage
-	go test -v -coverpkg=./... -coverprofile coverage/cover.out.tmp ./newserving/...
+	go test -v -coverpkg=./... -coverprofile coverage/cover.out.tmp ./serving/...
 
 test_runner:							## Requires ETCD to be installed and added to path
 	-mkdir coverage
@@ -308,7 +308,7 @@ containers: gen_grpc						## Build Docker containers for Minikube
 	minikube image build -f ./coordinator/Dockerfile . -t local/coordinator:stable & \
 	minikube image build -f ./metadata/Dockerfile . -t local/metadata:stable & \
 	minikube image build -f ./metadata/dashboard/Dockerfile . -t local/metadata-dashboard:stable & \
-	minikube image build -f ./newserving/Dockerfile . -t local/serving:stable & \
+	minikube image build -f ./serving/Dockerfile . -t local/serving:stable & \
 	minikube image build -f ./runner/Dockerfile . -t local/worker:stable & \
 	wait; \
 	echo "Build Complete"
